Rename `serializer-context`
